WIP Initial commit on Ember v1.12 > v1.13 upgrade
