+++ package.json
@@ -37,7 +37,7 @@
   "devDependencies": {
     "aws-sdk": "^2.3.3",
     "babel-eslint": "4.1.7",
-    "broccoli-asset-rev": "^2.0.0",
+    "broccoli-asset-rev": "^2.2.0",
     "broccoli-babel-transpiler": "5.1.0",
     "broccoli-concat": "0.0.12",
     "broccoli-es6modules": "^0.5.1",
@@ -50,18 +50,20 @@
     "broccoli-static-compiler": "^0.2.1",
     "broccoli-stew": "^0.3.1",
     "broccoli-wrap": "0.0.2",
-    "ember-cli": "0.2.0",
-    "ember-cli-app-version": "0.3.2",
-    "ember-cli-babel": "^5.0.0",
-    "ember-cli-content-security-policy": "0.3.0",
-    "ember-cli-dependency-checker": "0.0.8",
+    "ember-cli": "1.13.15",
+    "ember-cli-app-version": "^1.0.0",
+    "ember-cli-babel": "^5.1.5",
+    "ember-cli-content-security-policy": "0.4.0",
+    "ember-cli-dependency-checker": "^1.1.0",
+    "ember-disable-proxy-controllers": "^1.0.1",
     "ember-cli-eslint": "https://github.com/teddyzeenny/ember-cli-eslint#1.1.0",
-    "ember-cli-htmlbars": "0.7.4",
-    "ember-cli-ic-ajax": "0.1.1",
-    "ember-cli-inject-live-reload": "^1.3.0",
-    "ember-cli-qunit": "0.3.9",
-    "ember-cli-uglify": "1.0.1",
-    "ember-export-application-global": "^1.0.2",
+    "ember-cli-htmlbars": "^1.0.1",
+    "ember-cli-htmlbars-inline-precompile": "^0.3.1",
+    "ember-cli-ic-ajax": "0.2.4",
+    "ember-cli-inject-live-reload": "^1.3.1",
+    "ember-cli-qunit": "^1.0.4",
+    "ember-cli-uglify": "^1.2.0",
+    "ember-export-application-global": "^1.0.4",
     "ember-pikaday": "^0.2.1",
     "eslint-config-ember": "0.2.0",
     "fstream": "^1.0.8",
