+++ public/robots.txt
@@ -1,4 +1,3 @@
-# robotstxt.org/
-
+# http://www.robotstxt.org
 User-agent: *
-Disallow: /
+Disallow:
