+++ tests/.jshintrc
@@ -4,6 +4,7 @@
     "window",
     "location",
     "setTimeout",
+    "$",
     "-Promise",
     "define",
     "console",
@@ -25,7 +26,7 @@
   "node": false,
   "browser": false,
   "boss": true,
-  "curly": false,
+  "curly": true,
   "debug": false,
   "devel": false,
   "eqeqeq": true,
@@ -46,5 +47,6 @@
   "strict": false,
   "white": false,
   "eqnull": true,
-  "esnext": true
+  "esnext": true,
+  "unused": true
 }
