+++ tests/index.html
@@ -27,6 +27,7 @@
     <script src="assets/ember-inspector.js"></script>
     <script src="ember_debug.js"></script>
     <script src="testem.js"></script>
+    <script src="assets/tests.js"></script>
     <script src="assets/test-loader.js"></script>
 
     {{content-for 'body-footer'}}
